restore: make it possible to connect gofer objects with new containers
During restore, gofer sessions and inodes will need to re-establish
their relationships with the started gofer processes. Since we have
multiple containers, there has to be a way to tell to which container
the gofer object belong to.

The current session's connID is set to 9pfs-/mntpoint. If we change the
format of connID to: 9pfs-containerIndex/mntpoint, we can retrieve the
container index information from the gofer objects.

This made an assumption that all containers of a sandbox are started and
restored in exact the same order. e.g. on initial start, a root container
named rootbash is started using bash image, then a child container
containerfoo is started using ubuntu image and finally, yet another child
container containerbar are started using fedora image. Container rootbash's
index is 0(root container's index will always be 0), containerfoo's index
is 1 and containerbar's index is 2. On restore, three new containers will
be created and named: newrootbash, newcontainerfoo and newcontainerbar.
The three new containers shall be issued the restore cmd in exact this
order: first newrootbash(using the bash image), then newcontainerfoo(using
the ubuntu image) and finally newcontainerbashbar(using the fedora image),
so that gofer objects belonging to containerfoo will reconnect with the
gofer process of newcontainerfoo, etc.

Alternatively, we can use containerID in connID, but then we will have
to record the mapping of original containerID and the new containerID
somewhere. Image file is not a proper place because during checkpoint,
we have no idea of what the new containerIDs will be. specfile for the
to-be-restored container might be a proper place but then we will have
to record the original containerID in this specfile and parse this
information somehow and propagate the information to gofer object's
afterLoad() method. This feels a lot of complexity and isn't as clean as
the index method, although the index method enforce an restore order but
that can generally be fulfilled on deploy.

Suggestions are welcome.

Signed-off-by: Aaron Lu <ziqian.lzq@antfin.com>