Task: use index instead of string to identify a container
Restored container can have a different ID but the old containerID is
stored in Task sturcture. This causes a problem in that after restore,
when sending signal to a process in the restored container, the below
error will be printed:

$ sudo runsc --debug=true --debug-log=1.txt kill -pid 1 newcontainerfoo
failed to signal pid 1: signaling container "newcontainerfoo" PID 1:
signaling process in container "newcontainerfoo" PID 1: process 1 is
part of a different container: "containerfoo"

This problem is always there and isn't related to multi-container
restore support. I'm not sure if using different container IDs is a
supported use case for restore but I think it's a common use case
because docker-like tools will always generate a new ID for a newly
created container so it's quitely likely the restored container will
have a different ID compared to the original checkpointed container.

With this said, it might be possible to record a mapping of original
containerID to new containerID and during restore of Task, replace the
old containerID with the new one. The problem with this is already
stated in commit 158f0929c8b2("restore: make it possible to connect
gofer objects with new containers") so I chose to use container index to
identify a container. This will again enforce the rule that containers
inside a sandbox will have to be started and restored in exact the same
order.

TODO: we probably can change boot.execID to be also index based.

Signed-off-by: Aaron Lu <ziqian.lzq@antfin.com>